# frontend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Diverge API",
    "version": "1.0.0",
    "description": "Frozen HTTP API contract for Diverge financial narrative intelligence, sentiment tracking, multi-index calculation, integrity scoring, and reasoning trace explainability. This contract is frozen between backend (Antigravity) and frontend (Cursor). Any modifications must be called out explicitly."
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/auth/signup": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Register a new user account",
        "description": "Creates a new user profile and returns access & refresh tokens.",
        "operationId": "authSignup",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully with session tokens",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or email already registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Authenticate user",
        "description": "Validates user credentials and issues a new access token and refresh token pair.",
        "operationId": "authLogin",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentication successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Refresh expired access token",
        "description": "Exchanges a valid refresh token for a newly issued access token and refreshed refresh token.",
        "operationId": "authRefresh",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tokens refreshed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired refresh token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Retrieve authenticated user profile",
        "description": "Returns account details, permissions, and preferences for the currently authenticated bearer token session.",
        "operationId": "getAuthMe",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Authenticated user profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired authorization header",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Standard Bearer JWT Authorization header (e.g. 'Bearer <token>')"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": ["error_code", "message"],
        "properties": {
          "error_code": {
            "type": "string",
            "example": "INVALID_CREDENTIALS"
          },
          "message": {
            "type": "string",
            "example": "Invalid email or password provided."
          },
          "details": {
            "type": "object",
            "nullable": true,
            "additionalProperties": true,
            "example": {
              "field": "email"
            }
          }
        }
      },
      "SignupRequest": {
        "type": "object",
        "required": ["email", "password", "full_name"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "trader@example.com"
          },
          "password": {
            "type": "string",
            "format": "password",
            "minLength": 8,
            "example": "SecurePass123!"
          },
          "full_name": {
            "type": "string",
            "example": "Jane Doe"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "trader@example.com"
          },
          "password": {
            "type": "string",
            "format": "password",
            "example": "SecurePass123!"
          }
        }
      },
      "RefreshTokenRequest": {
        "type": "object",
        "required": ["refresh_token"],
        "properties": {
          "refresh_token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.refresh_token_payload"
          }
        }
      },
      "AuthTokenResponse": {
        "type": "object",
        "required": ["access_token", "refresh_token", "token_type", "expires_in"],
        "properties": {
          "access_token": {
            "type": "string",
            "description": "JWT Bearer access token for protected routes",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.access_token_payload"
          },
          "refresh_token": {
            "type": "string",
            "description": "Refresh token used to obtain a new access token",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.refresh_token_payload"
          },
          "token_type": {
            "type": "string",
            "example": "Bearer"
          },
          "expires_in": {
            "type": "integer",
            "description": "Access token validity in seconds",
            "example": 3600
          }
        }
      },
      "UserProfileResponse": {
        "type": "object",
        "required": ["id", "email", "full_name", "role", "created_at"],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "a1b2c3d4-e5f6-7a8b-9c0d-1e2f3a4b5c6d"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "trader@example.com"
          },
          "full_name": {
            "type": "string",
            "example": "Jane Doe"
          },
          "role": {
            "type": "string",
            "enum": ["standard", "pro", "admin"],
            "example": "standard"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-06-01T12:00:00.000Z"
          }
        }
      },
...
    }
  }
}